backend/requirements.txt
fastapi
uvicorn
pypdf
google-genai
python-dotenv
sentence-transformers
faiss-cpu
numpy
gunicorn
